move mongos specific upgrade into lib
