+++ lib/charms/mongos/v0/upgrade_helpers.py
@@ -1,3 +1,37 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Manager for handling Mongos in-place upgrades."""
+
+import abc
+import copy
+import enum
+import json
+import logging
+import pathlib
+import secrets
+import string
+import poetry.core.constraints.version as poetry_version
+from typing import Dict, List, Tuple, Optional
+
+from ops import BlockedStatus, MaintenanceStatus, StatusBase, Unit
+from ops.charm import CharmBase
+from ops.framework import Object
+from ops.model import ActiveStatus, BlockedStatus
+from tenacity import retry, stop_after_attempt, wait_fixed
+
+from charms.mongodb.v1.mongos import (
+    MongosConnection,
+)
+
+logger = logging.getLogger(__name__)
+
+
+WRITE_KEY = "write_value"
+ROLLBACK_INSTRUCTIONS = "To rollback, `juju refresh` to the previous revision"
+
+SHARD_NAME_INDEX = "_id"
+
 SHARD = "shard"
 PEER_RELATION_ENDPOINT_NAME = "upgrade-version-a"
 PRECHECK_ACTION_NAME = "pre-upgrade-check"
@@ -68,14 +102,14 @@ class AbstractUpgrade(abc.ABC):
     (https://docs.google.com/document/d/1tLjknwHudjcHs42nzPVBNkHs98XxAOT2BXGGpP7NyEU/)
     """
 
-    def __init__(self, charm_: ops.CharmBase) -> None:
+    def __init__(self, charm_: CharmBase) -> None:
         relations = charm_.model.relations[PEER_RELATION_ENDPOINT_NAME]
         if not relations:
             raise PeerRelationNotReady
         assert len(relations) == 1
         self._peer_relation = relations[0]
         self._charm = charm_
-        self._unit: ops.Unit = charm_.unit
+        self._unit: Unit = charm_.unit
         self._unit_databag = self._peer_relation.data[self._unit]
         self._app_databag = self._peer_relation.data[charm_.app]
         self._app_name = charm_.app.name
@@ -87,7 +121,7 @@ class AbstractUpgrade(abc.ABC):
             self._current_versions[version] = pathlib.Path(file_name).read_text().strip()
 
     @property
-    def unit_state(self) -> typing.Optional[UnitState]:
+    def unit_state(self) -> Optional[UnitState]:
         """Unit upgrade state."""
         if state := self._unit_databag.get("state"):
             return UnitState(state)
@@ -101,15 +135,13 @@ class AbstractUpgrade(abc.ABC):
         """Whether upgrade is supported from previous versions."""
         assert self.versions_set
         try:
-            previous_version_strs: typing.Dict[str, str] = json.loads(
-                self._app_databag["versions"]
-            )
+            previous_version_strs: Dict[str, str] = json.loads(self._app_databag["versions"])
         except KeyError as exception:
             logger.debug("`versions` missing from peer relation", exc_info=exception)
             return False
         # TODO charm versioning: remove `.split("+")` (which removes git hash before comparing)
         previous_version_strs["charm"] = previous_version_strs["charm"].split("+")[0]
-        previous_versions: typing.Dict[str, poetry_version.Version] = {
+        previous_versions: Dict[str, poetry_version.Version] = {
             key: poetry_version.Version.parse(value)
             for key, value in previous_version_strs.items()
         }
@@ -156,28 +188,26 @@ class AbstractUpgrade(abc.ABC):
         )
 
     @property
-    def _sorted_units(self) -> typing.List[ops.Unit]:
+    def _sorted_units(self) -> List[Unit]:
         """Units sorted from highest to lowest unit number."""
         return sorted((self._unit, *self._peer_relation.units), key=unit_number, reverse=True)
 
     @abc.abstractmethod
-    def _get_unit_healthy_status(self) -> ops.StatusBase:
+    def _get_unit_healthy_status(self) -> StatusBase:
         """Status shown during upgrade if unit is healthy."""
 
-    def get_unit_juju_status(self) -> typing.Optional[ops.StatusBase]:
+    def get_unit_juju_status(self) -> Optional[StatusBase]:
         """Unit upgrade status."""
         if self.in_progress:
             return self._get_unit_healthy_status()
 
     @property
-    def app_status(self) -> typing.Optional[ops.StatusBase]:
+    def app_status(self) -> Optional[StatusBase]:
         """App upgrade status."""
         if not self.in_progress:
             return
 
-        return ops.MaintenanceStatus(
-            "Upgrading. To rollback, `juju refresh` to the previous revision"
-        )
+        return MaintenanceStatus("Upgrading. To rollback, `juju refresh` to the previous revision")
 
     @property
     def versions_set(self) -> bool:
@@ -202,7 +232,7 @@ class AbstractUpgrade(abc.ABC):
 
     @property
     @abc.abstractmethod
-    def _unit_workload_container_versions(self) -> typing.Dict[str, str]:
+    def _unit_workload_container_versions(self) -> Dict[str, str]:
         """{Unit name: unique identifier for unit's workload container version}.
 
         If and only if this version changes, the workload will restart (during upgrade or
@@ -267,6 +297,103 @@ class AbstractUpgrade(abc.ABC):
 
         if not self.is_mongos_able_to_read_write():
             raise PrecheckFailed("mongos is not able to read/write.")
+
+
+class GenericMongoDBUpgrade(Object, abc.ABC):
+    """Substrate agnostif, abstract handler for upgrade events."""
+
+    def __init__(self, charm: CharmBase, *args, **kwargs):
+        super().__init__(charm, *args, **kwargs)
+        self._observe_events(charm)
+
+    @abc.abstractmethod
+    def _observe_events(self, charm: CharmBase) -> None:
+        """Handler that should register all event observers."""
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def _upgrade(self) -> AbstractUpgrade | None:
+        raise NotImplementedError()
+
+    def is_mongos_able_to_read_write(self) -> bool:
+        """Returns True if mongos is able to read and write."""
+        collection_name, write_value = self.get_random_write_and_collection()
+        logger.debug("-----\add_write_to_sharded_cluster\n----")
+        self.add_write_to_sharded_cluster(collection_name, write_value)
+
+        logger.debug("-----\nchecking write \n----")
+        write_replicated = self.confirm_excepted_write_cluster(
+            collection_name,
+            write_value,
+        )
+
+        self.clear_tmp_collection(collection_name)
+        if not write_replicated:
+            logger.debug("Test read/write to cluster failed.")
+            return False
+
+        return True
+
+    def get_random_write_and_collection(self) -> Tuple[str, str]:
+        """Returns a tuple for a random collection name and a unique write to add to it."""
+        choices = string.ascii_letters + string.digits
+        collection_name = "collection_" + "".join([secrets.choice(choices) for _ in range(32)])
+        write_value = "unique_write_" + "".join([secrets.choice(choices) for _ in range(16)])
+        return (collection_name, write_value)
+
+    def add_write_to_sharded_cluster(self, collection_name, write_value) -> None:
+        """Adds a the provided write to the provided database with the provided collection."""
+        with MongosConnection(self.charm.mongos_config) as mongos:
+            db = mongos.client[self.charm.database]
+            test_collection = db[collection_name]
+            write = {WRITE_KEY: write_value}
+            test_collection.insert_one(write)
+
+    @retry(
+        stop=stop_after_attempt(10),
+        wait=wait_fixed(1),
+        reraise=True,
+    )
+    def confirm_excepted_write_cluster(
+        self,
+        collection_name: str,
+        expected_write_value: str,
+    ) -> None:
+        """Returns True if the replica contains the expected write in the provided collection."""
+        with MongosConnection(self.charm.mongos_config) as mongos:
+            db = mongos.client[self.charm.database]
+            test_collection = db[collection_name]
+            query = test_collection.find({}, {WRITE_KEY: 1})
+            if query[0][WRITE_KEY] != expected_write_value:
+                return False
+
+        return True
+
+    def clear_tmp_collection(self, collection_name: str) -> None:
+        """Clears the temporary collection."""
+        with MongosConnection(self.charm.mongos_config) as mongos:
+            db = mongos.client[self.charm.database]
+            db.drop_collection(collection_name)
+
+    def _set_upgrade_status(self):
+        # In the future if we decide to support app statuses, we will need to handle this
+        # differently. Specifically ensuring that upgrade status for apps status has the lowest
+        # priority
+        if self.charm.unit.is_leader():
+            self.charm.app.status = self._upgrade.app_status or ActiveStatus()
+
+        # Set/clear upgrade unit status if no other unit status - upgrade status for units should
+        # have the lowest priority.
+        if isinstance(self.charm.unit.status, ActiveStatus) or (
+            isinstance(self.charm.unit.status, BlockedStatus)
+            and self.charm.unit.status.message.startswith(
+                "Rollback with `juju refresh`. Pre-upgrade check failed:"
+            )
+        ):
+            self.charm.status.set_and_share_status(
+                self._upgrade.get_unit_juju_status() or ActiveStatus()
+            )
 
 
 # END: Useful classes
