+++ src/upgrades/mongos_upgrade.py
@@ -63,9 +63,7 @@ class MongosUpgrade(Object):
         )
         self.framework.observe(charm.on.upgrade_charm, self._on_upgrade_charm)
 
-        self.framework.observe(
-            charm.on["force-upgrade"].action, self._on_force_upgrade_action
-        )
+        self.framework.observe(charm.on["force-upgrade"].action, self._on_force_upgrade_action)
         self.framework.observe(self.post_upgrade_event, self.run_post_upgrade_check)
 
     # BEGIN: Event handlers
@@ -150,9 +148,7 @@ class MongosUpgrade(Object):
         """Runs post-upgrade checks for after a shard/config-server/replset/cluster upgrade."""
         logger.debug("-----\nchecking mongos running\n----")
         if not self.charm.cluster.is_mongos_running():
-            logger.debug(
-                "Waiting for mongos router to be ready before finalising upgrade."
-            )
+            logger.debug("Waiting for mongos router to be ready before finalising upgrade.")
             event.defer()
             return
 
@@ -169,89 +165,6 @@ class MongosUpgrade(Object):
 
         logger.debug("upgrade of unit succeeded.")
         self._upgrade.unit_state = upgrade.UnitState.HEALTHY
-
-    def is_mongos_able_to_read_write(self) -> bool:
-        """Returns True if mongos is able to read and write."""
-        collection_name, write_value = self.get_random_write_and_collection()
-        logger.debug("-----\add_write_to_sharded_cluster\n----")
-        self.add_write_to_sharded_cluster(collection_name, write_value)
-
-        logger.debug("-----\nchecking write \n----")
-        write_replicated = self.confirm_excepted_write_cluster(
-            collection_name,
-            write_value,
-        )
-
-        self.clear_tmp_collection(collection_name)
-        if not write_replicated:
-            logger.debug("Test read/write to cluster failed.")
-            return False
-
-        return True
-
-    def get_random_write_and_collection(self) -> Tuple[str, str]:
-        """Returns a tuple for a random collection name and a unique write to add to it."""
-        choices = string.ascii_letters + string.digits
-        collection_name = "collection_" + "".join(
-            [secrets.choice(choices) for _ in range(32)]
-        )
-        write_value = "unique_write_" + "".join(
-            [secrets.choice(choices) for _ in range(16)]
-        )
-        return (collection_name, write_value)
-
-    def add_write_to_sharded_cluster(self, collection_name, write_value) -> None:
-        """Adds a the provided write to the provided database with the provided collection."""
-        with MongosConnection(self.charm.mongos_config) as mongos:
-            db = mongos.client[self.charm.database]
-            test_collection = db[collection_name]
-            write = {WRITE_KEY: write_value}
-            test_collection.insert_one(write)
-
-    @retry(
-        stop=stop_after_attempt(10),
-        wait=wait_fixed(1),
-        reraise=True,
-    )
-    def confirm_excepted_write_cluster(
-        self,
-        collection_name: str,
-        expected_write_value: str,
-    ) -> None:
-        """Returns True if the replica contains the expected write in the provided collection."""
-        with MongosConnection(self.charm.mongos_config) as mongos:
-            db = mongos.client[self.charm.database]
-            test_collection = db[collection_name]
-            query = test_collection.find({}, {WRITE_KEY: 1})
-            if query[0][WRITE_KEY] != expected_write_value:
-                return False
-
-        return True
-
-    def clear_tmp_collection(self, collection_name: str) -> None:
-        """Clears the temporary collection."""
-        with MongosConnection(self.charm.mongos_config) as mongos:
-            db = mongos.client[self.charm.database]
-            db.drop_collection(collection_name)
-
-    def _set_upgrade_status(self):
-        # In the future if we decide to support app statuses, we will need to handle this
-        # differently. Specifically ensuring that upgrade status for apps status has the lowest
-        # priority
-        if self.charm.unit.is_leader():
-            self.charm.app.status = self._upgrade.app_status or ActiveStatus()
-
-        # Set/clear upgrade unit status if no other unit status - upgrade status for units should
-        # have the lowest priority.
-        if isinstance(self.charm.unit.status, ActiveStatus) or (
-            isinstance(self.charm.unit.status, BlockedStatus)
-            and self.charm.unit.status.message.startswith(
-                "Rollback with `juju refresh`. Pre-upgrade check failed:"
-            )
-        ):
-            self.charm.status.set_and_share_status(
-                self._upgrade.get_unit_juju_status() or ActiveStatus()
-            )
 
     # END: helpers
